Reorganize the files
Create one dir per essay, so that each essay is independent and contains
everything: the sources plus the output.

Currently the two essays are written in LaTeX (so a PDF is needed for
the result), but I plan to write essays in other formats, depending on
my desires at the moment (it can be HTML, Plain TeX, ConTeXt, why not
Markdown, etc).